move Icon to seperate file to make room for more exports
