+++ packages/apollos-ui-kit/src/Icon/index.js
@@ -1,45 +1,3 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import { compose, pure } from 'recompose';
-import { flow, camelCase, upperFirst, kebabCase } from 'lodash';
-
-import Placeholder from '../Placeholder';
-
-import * as uikitIcons from '../theme/icons';
-
-const pascalCase = (string) =>
-  flow(
-    camelCase,
-    upperFirst
-  )(string);
-
-// Convenience component to render icons based on the icon's string name, like:
-// <Icon name="skip-next" {...allOtherPropsPassedToComponent} />
-//
-// Can also import the icon directly:
-// import { SkipNext } from '../theme/icons';
-// <SkipNext />
-
-const enhance = compose(pure);
-
-const Icon = enhance(({ name, size, isLoading = false, ...otherProps }) => {
-  const IconComponent = uikitIcons[pascalCase(name)];
-  return (
-    <Placeholder.Media size={size} hasRadius onReady={!isLoading}>
-      <IconComponent size={size} {...otherProps} />
-    </Placeholder.Media>
-  );
-});
-
-Icon.propTypes = {
-  name: PropTypes.oneOf(Object.keys(uikitIcons).map(kebabCase)).isRequired,
-  size: PropTypes.number,
-  fill: PropTypes.string,
-  isLoading: PropTypes.bool,
-};
-
-Icon.defaultProps = {
-  size: 32, // 32 is the default size used within the svg component
-};
+import Icon from './Icon';
 
 export default Icon;
